requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
Brotli==1.0.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
confluent-kafka==1.8.2
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
findspark==2.0.1
Flask==2.1.2
Flask-Compress==1.12
Flask-RESTful==0.3.9
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
kafka-python==2.0.2
lxml==4.8.0
MarkupSafe==2.1.1
multitasking==0.0.10
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
plotly==5.8.0
py4j==0.10.9.3
pymongo==4.1.1
pyrsistent==0.18.1
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.0.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yfinance==0.1.70
zipp==3.8.0
